chore: refactor screens and navigation interfaces to work with Jest
