+++ App.tsx
@@ -1,3 +1,4 @@
+import 'react-native-gesture-handler';
 import { NavigationContainer } from '@react-navigation/native';
 
 import { RootNavigator } from '~/navigation/root';
